fix: Handles Grain deactivation scenario in DummyStepLogic (#45)
Previously, `DummyStepLogic` spawns a long-running thread that runs independently of the Grain thread. When the long-running job completes, it needs to update the Step status in the Grain. Doing so when the Grain has been de-activated will cause an access violation.

The solution is to use DirectClient support in Orleans and let the long-running thread delegate schedule a Grain call to set its status. `_Call` on the `IStepGrain` interface is introduced for this purpose.

---------------------------------------------------
* handles long running job in DummyStepLogic

* Add timeout in TestObserver

* disable silo logging for tests

* clean up

* clean up StepGrainData